pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-crypto"
version = "0.1.0"
description = "Crypto Regime Analysis System — Modular, Agentic, Multi-Timeframe"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.1.0",
    "numpy>=1.25.0",
    "scipy>=1.11.0",
    "statsmodels>=0.14.0",
    "arch>=6.2.0",
    "pydantic>=2.5.0",
    "requests>=2.31.0",
    "backoff>=2.2.0",
    "pyyaml>=6.0",
    "matplotlib>=3.8.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
    "langchain-openai>=0.0.5",
    "python-telegram-bot>=20.7",
    "python-dotenv>=1.0.0",
    "pyarrow>=14.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "black>=23.10.0",
    "mypy>=1.6.0",
    "types-PyYAML>=6.0.0",
    "types-requests>=2.31.0",
]
pdf = [
    "markdown>=3.5.0",
    "weasyprint>=60.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

